- Implemented array panels into masterFrame - Improved the 'about' dialog - Scratched the idea of erasing by clicking the element it self     - Not enough time for proper error traping - Made timer work - Pause game now works - Have to figure out how to prompt to next stage properly - Have to figure out how to let the user know if they enterd a wrong answer

+++ src/Highscores.java
@@ -41,7 +41,7 @@ public class Highscores  extends HighscoresPanel{
      * @param difficulty The difficulty the player chose.
      * @param time The amount of time it took the player to finish the game.
      */
-    public Highscores (String name, int difficulty, int time) {
+    public Highscores (String name, int difficulty, long time) {
         for (int i = 0; i < players.size(); i++) {
             if (players.get(i).getTime() > time) {
                 players.add(i, new Player (name, difficulty, time));
